Feat: Migration to stacked architecture
