+++ lib/main.dart
@@ -1,20 +1,21 @@
 import 'package:flutter/material.dart';
-import 'package:mindspa_mobile/src/core/utilities/app_functions.dart';
-import 'package:mindspa_mobile/src/core/utilities/base_change_notifier.dart';
-import 'package:mindspa_mobile/src/features/startup/views/startup_view.dart';
-import 'package:mindspa_mobile/src/services/navigation_service.dart';
-import 'package:mindspa_mobile/src/services/snackbar_service.dart';
-import 'package:provider/provider.dart';
+import 'package:mindspa_mobile/src/app/app.locator.dart';
+import 'package:mindspa_mobile/src/app/app.router.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:mindspa_mobile/src/core/constant/app_strings.dart';
 import 'package:mindspa_mobile/src/core/theme.dart';
 
-import '../src/core/routes.dart';
+import 'package:mindspa_mobile/src/services/snackbar_service.dart';
+
+import 'package:stacked_services/stacked_services.dart';
+
+import '../src/UI/startup/startup_view.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   await Firebase.initializeApp();
+  setupLocator();
   runApp(const MyApp());
 }
 
@@ -23,30 +24,14 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [
-        ChangeNotifierProvider<SnackbarService>(
-          create: (context) => SnackbarService(),
-        ),
-        ChangeNotifierProvider<BaseChangeNotifier>(
-          create: (context) => BaseChangeNotifier(),
-        ),
-        Provider<NavigationService>(
-          create: (context) => NavigationService(),
-        ),
-        ChangeNotifierProvider<AppFunctions>(
-          create: (context) => AppFunctions(),
-        )
-      ],
-      builder: (context, _) => MaterialApp(
-        debugShowCheckedModeBanner: false,
-        title: AppStrings.mindSpa,
-        theme: AppTheme.lightTheme,
-        home: const StartupView(),
-        onGenerateRoute: Routes.generateRoute,
-        scaffoldMessengerKey:
-            context.read<SnackbarService>().scaffoldMessengerKey,
-      ),
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: AppStrings.mindSpa,
+      theme: AppTheme.lightTheme,
+      home: const StartupView(),
+      onGenerateRoute: StackedRouter().onGenerateRoute,
+      scaffoldMessengerKey: SnackbarServices.scaffoldMessengerKey,
+      navigatorKey: StackedService.navigatorKey,
     );
   }
 }
