+++ lib/src/UI/articles/articles_viewmodel.dart
@@ -1,0 +1,12 @@
+import 'package:mindspa_mobile/src/app/app.locator.dart';
+
+import 'package:stacked/stacked.dart';
+import 'package:stacked_services/stacked_services.dart';
+
+class ArticlesViewModel extends BaseViewModel {
+  final _navigationService = locator<NavigationService>();
+
+  void navigateBack() {
+    _navigationService.back();
+  }
+}
