+++ lib/src/UI/bottom_navigation/view/bottom_navigation_viewmodel.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class BottomNavigationViewModel extends IndexTrackingViewModel {}
