+++ lib/src/UI/for_you/for_you_viewmodel.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class ForYouViewModel extends BaseViewModel {}
