+++ lib/src/UI/mind_bot/view/mind_bot.dart
