+++ lib/src/UI/nutrition/nutrition_view.dart
@@ -1,0 +1,174 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:mindspa_mobile/src/UI/nutrition/nutrition_viewmodel.dart';
+import 'package:mindspa_mobile/src/core/constant/app_images.dart';
+import 'package:mindspa_mobile/src/core/constant/app_strings.dart';
+import 'package:mindspa_mobile/src/widgets/dummy_card.dart';
+import 'package:mindspa_mobile/src/widgets/spacing.dart';
+
+import 'package:stacked/stacked.dart';
+
+class NutritionView extends StatelessWidget {
+  const NutritionView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ViewModelBuilder<NutritionViewModel>.reactive(
+      viewModelBuilder: () => NutritionViewModel(),
+      builder: (
+        BuildContext context,
+        NutritionViewModel model,
+        Widget? child,
+      ) {
+        return Scaffold(
+          appBar: AppBar(
+            leading: IconButton(
+              onPressed: () => model.navigateback(),
+              icon: Icon(
+                Icons.arrow_back,
+                size: 30,
+                color: Theme.of(context).colorScheme.primaryVariant,
+              ),
+            ),
+            title: Text(
+              AppStrings.nutrition,
+              style: Theme.of(context)
+                  .textTheme
+                  .headline2!
+                  .copyWith(fontWeight: FontWeight.w100),
+            ),
+          ),
+          body: SingleChildScrollView(
+            child: Padding(
+              padding: const EdgeInsets.symmetric(
+                horizontal: 20.0,
+              ),
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                mainAxisAlignment: MainAxisAlignment.start,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  const Spacing.mediumHeight(),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Text(
+                        AppStrings.learnExperts,
+                        style: Theme.of(context)
+                            .textTheme
+                            .headline4!
+                            .copyWith(fontSize: 18),
+                      ),
+                      Text(
+                        AppStrings.seeAll,
+                        style: Theme.of(context)
+                            .textTheme
+                            .headline4!
+                            .copyWith(fontSize: 18),
+                      ),
+                    ],
+                  ),
+                  const Spacing.mediumHeight(),
+                  const DumbCards(),
+                  const Spacing.mediumHeight(),
+                  Text(
+                    AppStrings.explore,
+                    style: Theme.of(context)
+                        .textTheme
+                        .headline4!
+                        .copyWith(fontSize: 18),
+                  ),
+                  GridView(
+                    shrinkWrap: true,
+                    gridDelegate:
+                        const SliverGridDelegateWithFixedCrossAxisCount(
+                            crossAxisCount: 2),
+                    children: [
+                      for (var i = 0; i < 6; i++)
+                        Padding(
+                          padding: const EdgeInsets.symmetric(
+                              horizontal: 10.0, vertical: 10),
+                          child: Container(
+                            decoration: BoxDecoration(
+                              color: Theme.of(context).colorScheme.primary,
+                              borderRadius: const BorderRadius.all(
+                                Radius.circular(10),
+                              ),
+                            ),
+                            child: Card(
+                              child: Padding(
+                                padding: const EdgeInsets.all(20.0),
+                                child: Column(
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceEvenly,
+                                  children: [
+                                    FittedBox(
+                                      child: SvgPicture.asset(
+                                        AppImages.nutritionMain,
+                                        fit: BoxFit.contain,
+                                        height: 70,
+                                      ),
+                                    ),
+                                    const Spacing.mediumHeight(),
+                                    const Text(AppStrings.nutrition)
+                                  ],
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                    ],
+                  )
+                ],
+              ),
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
+
+
+
+//  GridView(
+//                   children: [
+//                     for (var i = 0; i < 10; i++)
+//                       Container(
+//                         margin: const EdgeInsets.symmetric(horizontal: 5),
+//                         decoration: BoxDecoration(
+//                             color: Theme.of(context).colorScheme.primary,
+//                             borderRadius:
+//                                 const BorderRadius.all(Radius.circular(10))),
+//                         child: Card(
+//                           child: Column(
+//                             mainAxisAlignment: MainAxisAlignment.spaceBetween,
+//                             children: [
+//                               SvgPicture.asset(
+//                                 AppImages.nutritionMain,
+//                                 height: 90,
+//                               ),
+//                               const Spacing.smallHeight(),
+//                               Container(
+//                                 height: 50,
+//                                 width: 150,
+//                                 color: Theme.of(context).colorScheme.onSurface,
+//                                 child: Center(
+//                                   child: Text(
+//                                     AppStrings.nutrition,
+//                                     style: Theme.of(context).textTheme.bodyText1,
+//                                   ),
+//                                 ),
+//                               )
+//                             ],
+//                           ),
+//                         ),
+//                       )
+//                   ],
+//                   shrinkWrap: true,
+//                   gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+//                     crossAxisCount: 2,
+//                     crossAxisSpacing: 10,
+//                     mainAxisSpacing: 10,
+//                   ),
+//                 )
