+++ lib/src/UI/onboarding/onboarding_viewmodel.dart
@@ -1,0 +1,12 @@
+import 'package:mindspa_mobile/src/app/app.locator.dart';
+import 'package:mindspa_mobile/src/app/app.router.dart';
+
+import 'package:stacked/stacked.dart';
+import 'package:stacked_services/stacked_services.dart';
+
+class OnboardingViewModel extends BaseViewModel {
+  final _navigationService = locator<NavigationService>();
+  void navigateback() => _navigationService.back();
+  void navigateToRegistration() =>
+      _navigationService.replaceWith(Routes.registerView);
+}
