+++ lib/src/UI/sleepandrelaxation/sleep_and_relaxation_viewmodel.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class SleepAndRelaxationViewModel extends BaseViewModel {}
