+++ lib/src/app/app.locator.dart
@@ -1,0 +1,27 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// StackedLocatorGenerator
+// **************************************************************************
+
+// ignore_for_file: public_member_api_docs
+
+import 'package:stacked/stacked.dart';
+import 'package:stacked/stacked_annotations.dart';
+import 'package:stacked_services/stacked_services.dart';
+
+import '../services/authentication_services.dart';
+import '../services/snackbar_service.dart';
+
+final locator = StackedLocator.instance;
+
+void setupLocator({String? environment, EnvironmentFilter? environmentFilter}) {
+// Register environments
+  locator.registerEnvironment(
+      environment: environment, environmentFilter: environmentFilter);
+
+// Register dependencies
+  locator.registerLazySingleton(() => NavigationService());
+  locator.registerLazySingleton(() => SnackbarServices());
+  locator.registerLazySingleton(() => AuthenticationServices());
+}
