+++ lib/src/core/constant/app_colors.dart
@@ -5,24 +5,13 @@ class AppColors {
   /// Its also has a private constructor so that the class cannot be instantiated outside the class.
   /// You can only access the static fields.
   AppColors._();
-  // static const Color green = Color(0XFF0E402D);
-  static const Color green = Color(0XFF1e412d);
-  static const Color lightGreen = Color(0XFF245D47);
-  static const Color lightOrange = Color(0XFFEA7C22);
+  static const Color green = Color(0XFF53796A);
+  static const Color lightGreen = Color(0XFF92AF96);
+  static const lightGreenVariant = Color(0XFF425F53);
   static const light = Color(0xFFFFFFFF);
-  static const dark = Color(0xFF000000);
-  static const grey = Color(0xFF8E8E8E);
-  static const weirdGrey = Color(0xFFE0E0E0);
+  static const grey = Color(0xFF999999);
   static const Color linksColor = Color(0XFF407afd);
   static const defaultShadow = BoxShadow(
     color: green,
   );
-  // static const Color kBodyColor = Color(0XFF8E8E8E);
-  // static const Color kIconWrapper = Color(0XFFC5DDC1);
-  // static const Color kGreenAccent = Color(0XFF135139);
-  //   static const Color kAccentColor = Color(0XFFEC8D3B);
-  // static const Color kLinksColor = Color(0XFF407afd);
-  // static const Color kBodyColor = Color(0XFF8E8E8E);
-  // static const Color kIconWrapper = Color(0XFFC5DDC1);
-  // static const Color kGreenAccent = Color(0XFF135139);
 }
