--- lib/src/features/home/notifier/home_notifier.dart
@@ -1,10 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-
-import '../../../core/utilities/base_change_notifier.dart';
-import '../../../repositories/authentication_repository.dart';
-
-class StartupNotifier extends BaseChangeNotifier {
-  final AuthenticationRepository authenticationRepository =
-      AuthenticationRepository();
-  User? get currentUser => authenticationRepository.loggedInUser;
-}
