--- lib/src/features/startup/notifiers/startup_notifier.dart
@@ -1,9 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-
-
-import '../../../core/utilities/base_change_notifier.dart';
-import '../../../repositories/authentication_repository.dart';
-
-class StartupNotifier extends BaseChangeNotifier {
-  User? get currentUser => AuthenticationRepository().loggedInUser;
-}
