+++ lib/src/services/snackbar_service.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:mindspa_mobile/src/core/constant/app_colors.dart';
 
-class SnackbarService with ChangeNotifier {
-  final scaffoldMessengerKey = GlobalKey<ScaffoldMessengerState>();
+class SnackbarServices {
+  static final scaffoldMessengerKey = GlobalKey<ScaffoldMessengerState>();
 
   ScaffoldFeatureController<SnackBar, SnackBarClosedReason> showSuccessSnackBar(
     String text,
