+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: mindspa_mobile
 description: A new Flutter project.
 
-publish_to: 'none'
+publish_to: "none"
 
 version: 1.0.0+1
 
@@ -9,22 +9,27 @@ environment:
   sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
+  animations: ^2.0.2
   cupertino_icons: ^1.0.2
   firebase_auth: ^3.3.5
   flutter:
     sdk: flutter
+  flutter_hooks: ^0.17.0
   flutter_native_splash: ^1.3.3
   flutter_svg: ^1.0.2
   google_fonts: ^2.2.0
   google_sign_in: ^5.2.3
   introduction_screen: ^2.1.0
-  logger: ^1.1.0
-  provider: ^6.0.2
+  stacked: ^2.2.8
+  stacked_hooks: ^0.2.1+1
+  stacked_services: ^0.8.16
 
 dev_dependencies:
+  build_runner: ^2.1.7
   flutter_lints: ^1.0.0
   flutter_test:
     sdk: flutter
+  stacked_generator: ^0.5.7
 
 flutter_native_splash:
   color: "#ffffff"
